+++ test/native/extconf.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require "mkmf"
+
+$CFLAGS << " -fobjc-arc"
+$LDFLAGS << " -framework Cocoa -framework Metal -framework QuartzCore"
+create_makefile("metaco_test")